refactor(daemon-commands): organize files in subdirectories and rename namespaces
